refactor(context): move Jinja2 template rendering to PullRequest class
